+++ src/dict_builder/divsufsort.rs
@@ -2945,7 +2945,7 @@ unsafe extern "C" fn sssort(
         PAi[0 as std::ffi::c_int
             as usize] = *PA
             .offset(*first.offset(-1_isize) as isize);
-        PAi[1 as usize] = n - 2;
+        PAi[1] = n - 2;
         a = first;
         i = *first.offset(-1_isize);
         while a < last
